Fix SwiftLint
